// packages/uikit/src/theme/types.ts
export enum SizeSpace {
    x000 = "x000",
    x025 = "x025",
    x050 = "x050",
    x075 = "x075",
    x100 = "x100",
    x150 = "x150",
    x200 = "x200",
    x250 = "x250",
    x300 = "x300",
    x400 = "x400",
    x500 = "x500",
    x600 = "x600",
    x700 = "x700",
    x800 = "x800",
}

export enum SizeText {
    title1 = "title1",
    title2 = "title2",
    title3 = "title3",
    headline = "headline",
    body = "body",
    callout = "callout",
    subheadline1 = "subheadline1",
    subheadline2 = "subheadline2",
    footnote = "footnote",
    caption1 = "caption1",
    caption2 = "caption2",
}

export enum ColorText {
    body = "body",
    accent = "accent",
    destructive = "destructive",
    button = "button",
    sectionHeader = "sectionHeader",
    subtitle = "subtitle",
    link = "link",
    hint = "hint",
}

export enum ColorBg {
    main = "main",
    secondary = "secondary",
    header = "header",
    section = "section",
    button = "button",
    sectionSeparator = "sectionSeparator",
}